update build test and cbdctrl test

+++ cbdctrl.py.data/cbdctrl.sh
@@ -325,6 +325,26 @@ function monitor_qemu() {
 }
 
 env
+cd ${kernel_dir}
+
+# List of variable names to check
+for var in ${config_list} ; do
+    # Check if the variable is set and whether it's true or false
+    if [ "${!var}" == "true" ]; then
+        ./scripts/config --enable "$var"
+    elif [ "${!var}" == "false" ]; then
+        ./scripts/config --disable "$var"
+    else
+        echo "Warning: $var is ${!var} not set to 'true' or 'false'"
+    fi
+done
+
+cat .config|grep CBD_
+
+make prepare; make -j 42 M=drivers/block/cbd/
+if [[ $? != 0 ]]; then
+	exit 1
+fi
 
 ssh ${blkdev_node} "rmmod cbd; insmod /workspace/linux_compile/drivers/block/cbd/cbd.ko"
 ssh ${backend_node} "rmmod cbd; insmod /workspace/linux_compile/drivers/block/cbd/cbd.ko"
@@ -420,7 +440,10 @@ if $multihost_mode; then
 	kill_qemu_pid=$!
 fi
 
-run_remote_cmd $blkdev_node "/root/run_xfstests.sh"
+run_remote_cmd $blkdev_node "cd /root/xfstests/;./check generic/001"
+if [[ $? != 0 ]]; then
+	exit 1
+fi
 
 if $multihost_mode; then
 	kill $kill_qemu_pid
@@ -449,3 +472,4 @@ if $multihost_mode; then
 fi
 # Print PASS if the command was successful
 echo "PASS"
+exit 0
